Steps 6-9 (Parts 4.18-4.21)

+++ part4/server/controllers/blogs.js
@@ -1,23 +1,38 @@
 const express = require('express')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-const router = express.Router()
+const blogsRouter = express.Router()
 
-router.get('/', async (req, res) => {
+blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   res.json(blogs)
 })
 
-router.post('/', async (req, res) => {
-  const user = await User.findOne()
+blogsRouter.post('/', async (req, res) => {
+  const body = req.body
+  console.log('Token before verification: ', req.token) // Debugging
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
   const blog = new Blog({
-    ...req.body,
-    user: user._id
+    url     :  body.url,
+    title   :  body.title,
+    author  :  body.author,
+    user    :  user._id,
+    likes   :  body.likes || 0
   })
 
   if (!blog.title || !blog.url) {
-    return res.status(400).end()
+    return res
+      .status(400)
+      .json({ error: 'title and url are required' })
+      .end()
   }
 
   const savedBlog = await blog.save()
@@ -26,12 +41,25 @@ router.post('/', async (req, res) => {
   res.status(201).json(savedBlog)
 })
 
-router.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'invalid token' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return res.status(401).json({ error: 'only the creator can delete blogs' })
+  }
+
   await Blog.findByIdAndRemove(req.params.id)
   res.status(204).end()
 })
 
-router.put('/:id', async (req, res) => {
+
+blogsRouter.put('/:id', async (req, res) => {
   const body = req.body
   const blog = {
     title   :  body.title,
@@ -45,4 +73,4 @@ router.put('/:id', async (req, res) => {
   }
 })
 
-module.exports = router
+module.exports = blogsRouter
